Add optional client_nonce for OIDC logins
This optional nonce hardens the multi-step OIDC login process. If the
client provides a nonce at the start (auth_url) phase, it must provide
the same nonce as part of the callback.

The motivation for this is primarily CLI logins. The CLI listener is
necessarily unencrypted, and theoretically if an attacker stole the OAuth
authorization_code enroute to the CLI listener, it could post that to
Vault and gain access. The CLI leverages the new client_nonce support to
harden this flow. Only the CLI process (and Vault) will know the
client_nonce, so an attacker will be unable to complete the login even
if they manage to steal the authorization_code.