Select browser using environment variable
Currently `vault login` always opens the default browser for
login via OIDC. However for security reasons there might be the
need to use a different browser which is logged in as a higher
privileged user (you might not want to browse as an admin
for regular work..).

This PR adds the option to select a different browser by
setting the `BROWSER` environment variable. Example:
`BROWSER=/usr/bin/firefox vault login -method=oidc`.
The default behaviour will still be the same.