Add a -no-launch argument
This change allows a user to opt-out of the auto-launching of their
default browser which may already have a logged in OIDC session. Since
SSO can be a bit tricky to logout/switch users, this provides the user
an escape option for more careful handling of the login attempt.

Primary use-case is Firefox Containers or users using separate browser
profiles to manage multiple logins.